Corrected r12662 where a guild break would trigger the OnGuildBreak event for all castles (even of other guilds). Corrected some x64 compilation warnings in ers/malloc files. cleaned up a case of bad variable reuse.
git-svn-id: https://svn.code.sf.net/p/rathena/svn/trunk@12666 54d463be-8e91-2dee-dedb-b68131a5f0ec

+++ src/common/ers.c
@@ -245,7 +245,7 @@ static void ers_obj_free_entry(ERS self, void *entry)
  * @see ERS_impl#size
  * @see ERS_impl::vtable#entry_size
  */
-static uint32 ers_obj_entry_size(ERS self)
+static size_t ers_obj_entry_size(ERS self)
 {
 	ERS_impl obj = (ERS_impl)self;
