+++ src/common/ers.h
@@ -103,7 +103,7 @@ typedef struct eri {
 	 * @param self Interface of the entry manager
 	 * @return Size of the entries of this manager in bytes
 	 */
-	uint32 (*entry_size)(struct eri *self);
+	size_t (*entry_size)(struct eri *self);
 
 	/**
 	 * Destroy this instance of the manager.
@@ -120,7 +120,7 @@ typedef struct eri {
 // Use memory manager to allocate/free and disable other interface functions
 #	define ers_alloc(obj,type) (type *)aMalloc(sizeof(type))
 #	define ers_free(obj,entry) aFree(entry)
-#	define ers_entry_size(obj) (uint32)0
+#	define ers_entry_size(obj) (size_t)0
 #	define ers_destroy(obj)
 // Disable the public functions
 #	define ers_new(size) NULL
